+++ src/main/java/com/codetaylor/mc/pyrotech/modules/core/block/BlockOreDenseQuartzRocks.java
@@ -1,0 +1,61 @@
+package com.codetaylor.mc.pyrotech.modules.core.block;
+
+import com.codetaylor.mc.pyrotech.modules.core.block.spi.BlockOreDenseQuartzBase;
+import net.minecraft.block.state.IBlockState;
+import net.minecraft.util.math.AxisAlignedBB;
+import net.minecraft.util.math.BlockPos;
+import net.minecraft.world.IBlockAccess;
+
+import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
+import java.util.Random;
+
+public class BlockOreDenseQuartzRocks
+    extends BlockOreDenseQuartzBase {
+
+  public static final String NAME = "dense_quartz_ore_rocks";
+
+  private static final AxisAlignedBB AABB = new AxisAlignedBB(0, 0, 0, 1, 0.0625, 1);
+
+  @Override
+  public String getUnlocalizedName() {
+
+    return "tile.pyrotech.dense_quartz_ore";
+  }
+
+  @Override
+  public int quantityDropped(Random random) {
+
+    return 1 + random.nextInt(2);
+  }
+
+  @Override
+  public int getExpDrop(IBlockState state, IBlockAccess world, BlockPos pos, int fortune) {
+
+    return 1 + RANDOM.nextInt(2);
+  }
+
+  // ---------------------------------------------------------------------------
+  // - Collision
+  // ---------------------------------------------------------------------------
+
+  @Nonnull
+  @Override
+  public AxisAlignedBB getBoundingBox(IBlockState state, IBlockAccess source, BlockPos pos) {
+
+    return AABB;
+  }
+
+  @Nullable
+  @Override
+  public AxisAlignedBB getCollisionBoundingBox(IBlockState blockState, @Nonnull IBlockAccess world, @Nonnull BlockPos pos) {
+
+    return NULL_AABB;
+  }
+
+  @Override
+  public boolean isPassable(IBlockAccess world, BlockPos pos) {
+
+    return true;
+  }
+}